Bump minimum-supported Gradle version to 9
This allows us to use Kotlin 2.2 and Java 17. The forthcoming Kotlin 2.3.0 will no longer support compiling to 1.8, so we must bump.
