+++ README.md
@@ -50,9 +50,10 @@ apply plugin: 'com.jakewharton.test-distribution'
 This plugin relies on Gradle internal APIs.
 As such, certain versions of the plugin only work with certain versions of Gradle.
 
-| Gradle       | Plugin        |
-|--------------|---------------|
-| 8.10 - 9.1.0 | 0.1.0 - 0.2.0 |
+| Gradle        | Plugin                 |
+|---------------|------------------------|
+| 9.0.0 - 9.1.0 | 0.2.0 - 0.3.0-SNAPSHOT |
+| 8.10 - 8.14.3 | 0.1.0 - 0.2.0          |
 
 Gradle versions newer than those listed may be supported, but have not been tested.
